requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mlflow
jupyter
pyarrow
fastparquet
dvc
pathspec>=0.10.3
pyyaml
optuna==4.2.1
hydra-core==1.3.2
omegaconf==2.3.0
joblib
pytest>=8.0
black>=24.0
flake8>=7.0
